+++ uninstall.php
@@ -12,6 +12,8 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
 	exit;
 }
 
+// Nonce?: https://core.trac.wordpress.org/ticket/38661
+
 if (isset($_REQUEST['plugin']) && $_REQUEST['plugin'] != 'htmx-api-wp/htmx-api-wp.php' && $_REQUEST['action'] != 'delete-plugin') {
 	wp_die('Error uninstalling: wrong plugin.');
 }